rename: 이벤트 도메인에 맞게 파일명 수정
